Add external configury boilerplate
HTSlib provides macros in ax_with_htslib.m4 that add a --with-htslib
configure option for specifying an HTSlib to build against and checking
that it is usable. Eventually this should be available from a contrib/
directory in HTSlib releases and/or the Autoconf Archive. This copy is
copied from SAMtools.

Providing install-sh is a prerequisite for AC_CONFIG_SUBDIRS, which
AX_WITH_HTSLIB uses. This is the current install-sh from Gnulib.