Add Makefile.in and configure.ac
The configure script uses AX_WITH_HTSLIB to find an HTSlib installation
or source tree. The Makefile uses the HTSLIB_* output variables to add
appropriate preprocessing and linking options.

To link against a static libhts.a, this requires HTSlib to have been
built with --disable-lzma --disable-bz2 --disable-libcurl (or for
tinycount to be built with additional externally-supplied LIBS entries)
as at present AX_WITH_HTSLIB does not account for these additional
HTSlib-imposed dependencies.