ax_with_htslib.m4: Support separate HTSlib build directory
Also recognise an HTSlib separate build directory, and treat it the
same as a source directory -- as for external projects' purposes this
is identical apart from -I needing to point to the associated source
directory (only).

Pick HTSSRCDIR out of the HTSlib build directory's htslib_vars.mk
(generated from builddirs_vars.mk.in). As for other $ac_cv_htslib_which
categories, @HTSLIB_CPPFLAGS@ and @HTSLIB_LDFLAGS@ need to be fully
expanded: '$(HTSDIR)' would not be available for external projects
that don't use @HTSDIR@ or htslib.mk.